backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
pgvector==0.2.5
redis==5.0.3
celery==5.3.6
python-dotenv==1.0.1
langchain==0.1.16
langchain-openai==0.1.3
langchain-google-genai==1.0.1
httpx==0.27.0
pytest==8.1.1
